update login and register

+++ src/constants/auth.js
@@ -11,22 +11,40 @@ export const fakeAuthProvider = {
   register(user, callback) {
     // eslint-disable-next-line no-mixed-operators
     const users = (localStorage.getItem("users") && JSON.parse(localStorage.getItem("users"))) || []
-    users.push(user)
-    localStorage.setItem('users', JSON.stringify(users))
+    let message = ''
+    if (users.length > 0) {
+      message = 'User is existed'
+      // not allow existed user
+      const checkExistUser = users.find((item) => item.username === user.username)
+      if (!checkExistUser) {
+        message = 'Register user successfully'
+        fakeAuthProvider.isAuthenticated = true
+        users.push(user)
+        localStorage.setItem('users', JSON.stringify(users))
+      }
+    } else {
+      message = 'Register user successfully'
+      users.push(user)
+      localStorage.setItem('users', JSON.stringify(users))
+    }
 
-    setTimeout(callback(true), 100) // fake async
+    setTimeout(callback(fakeAuthProvider.isAuthenticated, message), 100) // fake async
   },
   signin(user, callback) {
     // eslint-disable-next-line no-mixed-operators
     const users = (localStorage.getItem("users") && JSON.parse(localStorage.getItem("users"))) || []
+    let message = ''
     if (users.length > 0) {
+      message = 'User not found'
       const checkExistUser = users.find((item) => item.username === user.username && item.password === user.password)
       if (checkExistUser) {
+        message = 'Login Successfully'
         fakeAuthProvider.isAuthenticated = true
+        localStorage.setItem("user", JSON.stringify(checkExistUser))
       }
     }
 
-    setTimeout(callback(fakeAuthProvider.isAuthenticated), 100) // fake async
+    setTimeout(callback(fakeAuthProvider.isAuthenticated, message), 100) // fake async
   },
   signout(callback) {
     fakeAuthProvider.isAuthenticated = false
